+++ src/test/java/io/github/pitzzahh/atm/dao/InDatabaseAtmServiceTest.java
@@ -1,4 +1,4 @@
-package io.github.pitzzahh.service;
+package io.github.pitzzahh.atm.dao;
 
 import java.time.*;
 import java.util.Map;
