+++ src/test/java/io/github/pitzzahh/atm/dao/InMemoryAtmServiceTest.java
@@ -1,11 +1,11 @@
-package io.github.pitzzahh.atm.database;
+package io.github.pitzzahh.atm.dao;
 
+import io.github.pitzzahh.atm.dao.InMemory;
 import io.github.pitzzahh.atm.service.AtmService;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.TestMethodOrder;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.BeforeEach;
-import static org.junit.jupiter.api.Assertions.*;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 class InMemoryAtmServiceTest {
